# FIX Prevent unnecessary call to config system which doesn't exist yet