@micmania1 2015-03-03
Before:

	Transactions:                      50 hits
	Availability:                 100.00 %
	Elapsed time:                  76.32 secs
	Data transferred:               0.07 MB
	Response time:                  6.61 secs
	Transaction rate:               0.66 trans/sec
	Throughput:                     0.00 MB/sec
	Concurrency:                    4.33
	Successful transactions:          50
	Failed transactions:               0
	Longest transaction:           10.84
	Shortest transaction:           2.76


After:

	Transactions:                      50 hits
	Availability:                 100.00 %
	Elapsed time:                  71.81 secs
	Data transferred:               0.07 MB
	Response time:                  6.20 secs
	Transaction rate:               0.70 trans/sec
	Throughput:                     0.00 MB/sec
	Concurrency:                    4.32
	Successful transactions:          50
	Failed transactions:               0
	Longest transaction:            8.89
	Shortest transaction:           2.78