@dhensby 2015-03-03
It'd be interesting to see how significant the savings are once #3934 is merged