@dhensby 2015-03-03
Ok, after merging in the config_lru changes and commenting out line 88 from core.php I'm seeing improvements of about 4.5%