@dhensby 2015-03-03
My point about the Config_LRU is that the impact of this fix will be less once the LRU changes are in because the slow LRU currently impacts this call