@dhensby 2015-03-03
Well, considering this doesn't seem to break anything, I'd be happy to take it in.