@nyeholt 2015-03-03
The 'dependencies' static declaration overlaps a fair amount with the config based 'properties' definition, but exists as a bit of syntactic sugar to avoid the need to explicitly declare those via config, and provides the inherited lookup of those. If that call to 'dependencies' is causing slowness, then I'd love to get rid of it :D Of course, that would be a backwards compatible breaker so release version would need to be later.