@micmania1 2015-03-03
@nyeholt its not so much an issue with the dependencies static, its just that in this case its being called so early that the configuration has not yet loaded so its impossible to find a value for it. However, it continues to look for it which is quite expensive.