@tractorcow 2015-03-05
I'm going to trust that the work you did and benchmarking demonstrates this is an improvement. I think the best course of action is to put this out in the wild and see if any other issues arise.