@nyeholt 2015-03-05
This will be BC breaking for anyone that currently relies on binding the config in based on its existence as a named service; however, the likelihood of that being the case is very low given most people refer to the config via a direct static reference anyway.